# pkg/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Auto Generate Mock-up Data",
        "title": "Generate Mock-up API",
        "contact": {
            "name": "HEEPOKE Support",
            "url": "https://github.com/HEEPOKE",
            "email": "Damon1FX@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:6476",
    "basePath": "/apis",
    "paths": {
        "/generate/get-details": {
            "get": {
                "description": "Get List All Generate Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Generate"
                ],
                "summary": "Get List All Generate Data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/examples.SuccessGenerateGetAllResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/examples.FailedCommonResponse"
                        }
                    }
                }
            }
        },
        "/generate/mockup-data": {
            "post": {
                "description": "Get Auto Generate Mock up data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Generate"
                ],
                "summary": "Get Auto Generate Mock up data",
                "parameters": [
                    {
                        "description": "GenerateRequestBody",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/examples.GenerateExample"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/examples.SuccessCheckDataFromKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/examples.FailedCommonResponse"
                        }
                    }
                }
            }
        },
        "/utilities/check-data": {
            "get": {
                "description": "Get Data From Key in Cache",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Utilities"
                ],
                "summary": "Get Data From Key in Cache",
                "parameters": [
                    {
                        "type": "string",
                        "example": "123456",
                        "description": "key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/examples.SuccessCheckDataFromKeyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/examples.FailedCommonResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "enums.Category": {
            "type": "string",
            "enum": [
                "name",
                "firstname",
                "lastname",
                "email",
                "tel",
                "password",
                "word",
                "sentence",
                "bool",
                "int",
                "int8",
                "int32",
                "int16",
                "int64",
                "uint",
                "uint8",
                "uint16",
                "uint32",
                "uint64",
                "float32",
                "float64",
                "byte",
                "rune"
            ],
            "x-enum-varnames": [
                "NAME",
                "FIRSTNAME",
                "LASTNAME",
                "EMAIL",
                "TEL",
                "PASSWORD",
                "WORD",
                "SENTENCE",
                "BOOL",
                "INT",
                "INT8",
                "INT32",
                "INT16",
                "INT64",
                "UINT",
                "UINT8",
                "UINT16",
                "UINT32",
                "UINT64",
                "FLOAT32",
                "FLOAT64",
                "BYTE",
                "RUNE"
            ]
        },
        "examples.FailedCommonResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "$ref": "#/definitions/examples.FailedStatusMessage"
                }
            }
        },
        "examples.FailedStatusMessage": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "0001"
                },
                "description": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "Failed"
                },
                "service": {
                    "type": "string"
                }
            }
        },
        "examples.GenerateExample": {
            "type": "object",
            "required": [
                "columns",
                "quantity",
                "table"
            ],
            "properties": {
                "columns": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/examples.GenerateOptions"
                    }
                },
                "key": {
                    "type": "string",
                    "example": ""
                },
                "quantity": {
                    "type": "integer",
                    "example": 1000
                },
                "table": {
                    "type": "string",
                    "example": "users"
                }
            }
        },
        "examples.GenerateOptions": {
            "type": "object",
            "properties": {
                "default": {
                    "type": "string",
                    "example": ""
                },
                "gen": {
                    "type": "boolean"
                },
                "length": {
                    "type": "integer"
                },
                "types": {
                    "$ref": "#/definitions/enums.Category"
                }
            }
        },
        "examples.GenerateResponseExample": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2022-01-01 00:00:00"
                },
                "deleted_at": {
                    "type": "string",
                    "example": "0001-01-01 00:00:00"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "key": {
                    "type": "string",
                    "example": "123456"
                },
                "quantity": {
                    "type": "integer",
                    "example": 1000
                },
                "table": {
                    "type": "string",
                    "example": "users"
                },
                "time_expired": {
                    "type": "string",
                    "example": "2022-01-01 00:00:00"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2022-01-01 00:00:00"
                }
            }
        },
        "examples.SuccessCheckDataFromKeyResponse": {
            "type": "object",
            "properties": {
                "payload": {
                    "$ref": "#/definitions/models.JsonStructure"
                },
                "status": {
                    "$ref": "#/definitions/examples.SuccessStatusMessage"
                }
            }
        },
        "examples.SuccessGenerateGetAllResponse": {
            "type": "object",
            "properties": {
                "payload": {
                    "$ref": "#/definitions/examples.GenerateResponseExample"
                },
                "status": {
                    "$ref": "#/definitions/examples.SuccessStatusMessage"
                }
            }
        },
        "examples.SuccessStatusMessage": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "0000"
                },
                "description": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "service": {
                    "type": "string"
                }
            }
        },
        "models.JsonStructure": {
            "type": "object",
            "properties": {
                "datas": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": true
                    }
                },
                "key": {
                    "type": "string",
                    "example": "123456"
                },
                "quantity": {
                    "type": "integer",
                    "example": 1000
                },
                "table": {
                    "type": "string",
                    "example": "users"
                }
            }
        }
    }
}